Refactor environment variables handling
