+++ src/api/routes/user.js
@@ -1,5 +1,5 @@
-const { authJwt } = require("../middleware");
-const controller = require("../controllers/user");
+const { authJwt } = require("../../middleware");
+const controller = require("../../controllers/user");
 
 module.exports = function(app) {
   app.use(function(req, res, next) {
