--- src/config/auth.js
@@ -1,3 +1,0 @@
-module.exports = {
-  secret: process.env.JWT_SECRET
-};
