--- src/config/database.js
@@ -1,13 +1,0 @@
-module.exports = {
-  HOST: 'localhost',
-  USER: process.env.DB_USER,
-  PASSWORD: process.env.DB_PASS,
-  DB: process.env.DB_NAME,
-  dialect: "postgres",
-  pool: {
-    max: 5,
-    min: 0,
-    acquire: 30000,
-    idle: 10000
-  }
-};
